+++ CHANGELOG.md
@@ -7,6 +7,7 @@ y este proyecto adhiere a [Semantic Versioning]( https://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Menú hamburguesa con transición de altura.
+- Agregado de tags <article>.
 
 ### Changed
 -